requirements-sync.txt
supabase
requests
python-dotenv

